refactor : 중복되는 기능 제거 및 정리
