+++ README.md
@@ -35,7 +35,9 @@
   - [x] 당첨 번호와 중복되지 않는 1 ~ 45 중 하나는 보너스 볼이다.
   - [x] 당첨 번호와 일치하는 로또 번호 개수를 반환한다.
 
-- [x] 로또 통계
+- [x] 로또 상금
   - [x] 로또 티켓의 당첨 순위에 따른 당첨 금액을 반환한다.
-  - [x] 당첨 번호와 일치하는 로또 티켓 개수를 반환한다.
+  
+- [x] 로또 통계
+  - [] 당첨 번호 개수가 같은 로또 티켓 개수를 반환한다.
   - [x] 당첨 순위의 수익률을 계산한다.
