+++ src/test/java/lotto/WinningStatisticsTest.java
@@ -1,7 +1,6 @@
 package lotto;
 
 import lotto.domain.LottoPrize;
-import lotto.domain.LottoTicket;
 import lotto.domain.WinningNumbers;
 import lotto.domain.WinningStatistics;
 import org.junit.jupiter.api.DisplayName;
@@ -16,24 +15,13 @@ import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class WinningStatisticsTest {
+    private static WinningNumbers winningNumbers = generateWinningNumber();
+    private static WinningStatistics winningStatistics = new WinningStatistics(winningNumbers);
 
-    @ParameterizedTest
-    @MethodSource("generateLottoTicket")
-    @DisplayName("당첨 번호와 일치하는 로또 티켓 개수를 반환한다")
-    void winning_number_match_count_statistics(int[] numbers, int expectedCount) {
-        //given
-        LottoTicket lottoTicket = new LottoTicket(numbers);
-
+    private static WinningNumbers generateWinningNumber() {
         List<Integer> inputNumbers = Arrays.asList(1, 2, 3, 4, 5, 6);
         int bonusNumber = 45;
-        WinningNumbers winningNumbers = new WinningNumbers(inputNumbers, bonusNumber);
-
-        //when
-        WinningStatistics winningStatistics = new WinningStatistics();
-        int matchedCount = winningStatistics.matchedCount(lottoTicket, winningNumbers);
-
-        //then
-        assertThat(matchedCount).isEqualTo(expectedCount);
+        return new WinningNumbers(inputNumbers, bonusNumber);
     }
 
     @ParameterizedTest
@@ -46,7 +34,6 @@ public class WinningStatisticsTest {
                 .sum();
 
         //when
-        WinningStatistics winningStatistics = new WinningStatistics();
         float profit = winningStatistics.profitRate(matchedCounts.length, totalPrize);
 
         //then
@@ -62,18 +49,6 @@ public class WinningStatisticsTest {
                 Arguments.of(new int[]{0, 1, 0, 0, 5}, false, 300.0f),
                 Arguments.of(new int[]{0, 1, 0, 0, 5}, true, 6000.0f),
                 Arguments.of(new int[]{0, 1, 0, 0, 6}, false, 400000.0f)
-        );
-    }
-
-    private static Stream<Arguments> generateLottoTicket() {
-        return Stream.of(
-                Arguments.of(new int[]{1, 2, 3, 4, 5, 6}, 6),
-                Arguments.of(new int[]{1, 2, 3, 4, 5, 7}, 5),
-                Arguments.of(new int[]{1, 2, 3, 4, 7, 8}, 4),
-                Arguments.of(new int[]{1, 2, 3, 7, 8, 9}, 3),
-                Arguments.of(new int[]{1, 2, 7, 8, 9, 10}, 2),
-                Arguments.of(new int[]{1, 7, 8, 9, 10, 11}, 1),
-                Arguments.of(new int[]{8, 9, 10, 11, 12, 13}, 0)
         );
     }
 }
